Refactor: 패키지 및 롬복 게터 리팩터링
