+++ src/main/java/kdt/prgms/springbootboard/dto/PostDto.java
@@ -1,5 +1,11 @@
 package kdt.prgms.springbootboard.dto;
 
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 public class PostDto {
 
     private Long id;
@@ -18,38 +24,6 @@ public class PostDto {
         this.id = id;
         this.title = title;
         this.content = content;
-        this.userDto = userDto;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public UserDto getUserDto() {
-        return userDto;
-    }
-
-    public void setUserDto(UserDto userDto) {
         this.userDto = userDto;
     }
 }
